Add rexml dependency to fix ruby-head builds
The rexml library is no longer bundled with ruby in ruby-head; it will
be distributed as a separate gem. The crack gem, used by webmock, uses
rexml but has not been updated to reflect the rexml dependency.

Until crack is updated, work around this issue by explicitly adding the
rexml dependency to this project.

Fixes this error:

```
$ bundle exec rake
File does not exist: rexml/parsers/streamparser
rake aborted!
```